refactor(language-selection-screen): PR changes
