+++ app/constants/languages.ts
@@ -1,4 +1,4 @@
-import { IRadioButtonListItem } from '../../../../components';
+import { IRadioButtonListItem } from '../components';
 
 export const LANGUAGES: ReadonlyArray<IRadioButtonListItem> = [
   { id: 'en', title: 'languages.english' },
