--- app/screens/Onboarding/Parents/LanguageSelection/LanguageSelection.data.ts
@@ -1,7 +1,0 @@
-import { IRadioButtonListItem } from '../../../../components';
-
-export const LANGUAGES: ReadonlyArray<IRadioButtonListItem> = [
-  { id: 'en', title: 'languages.english' },
-  { id: 'uk', title: 'languages.ukrainian' },
-  { id: 'rus', title: 'languages.russian' },
-];
